Changes per https://github.com/DoSomething/gambit-admin/pull/83#pullrequestreview-182336894
